changed int numDigits to a double and added a (long) in front of the second n

+++ c_version/kaprekarUtils.c
@@ -32,9 +32,9 @@ int isKaprekar(int n) {
 
   int i;
   //compute the square
-  long square = n * n;
+  long square = n * (long) n;
   //compute the number of digits of the square
-  int numDigits = log10(square) + 1;
+  double numDigits = log10(square) + 1;
   long modulus = 1;
   long first, second;
